テーブルコンポーネント追加
