+++ front/.storybook/main.js
@@ -4,12 +4,13 @@ module.exports = {
   webpackFinal:async(config)=>{
     config.resolve.alias={
       ...config.resolve.alias,
-      "@components":path.resolve(__dirname,"../components")
+      "@components":path.resolve(__dirname,"../components"),
+      "@utils":path.resolve(__dirname,"../utils"),
+      "@assets":path.resolve(__dirname,"../public")
     }
     return config
   },
   "stories": [
-    "../stories/**/*.stories.@(js|jsx|ts|tsx)",
     "../components/**/*.stories.@(js|jsx|ts|tsx)"
   ],
   "addons": [
@@ -18,6 +19,7 @@ module.exports = {
     "@storybook/addon-interactions",
     "@storybook/addon-postcss"
   ],
+  "staticDirs":["../public"],
   "framework": "@storybook/react",
   "core": {
     "builder": "@storybook/builder-webpack5"
