+++ front/.storybook/preview.js
@@ -1,4 +1,10 @@
 import "../styles/globals.css"
+import * as nextImage from 'next/image';
+
+Object.defineProperty(nextImage, 'default', {
+  configurable: true,
+  value: props => <img {...props} />
+});
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
@@ -9,3 +15,11 @@ export const parameters = {
     },
   },
 }
+
+Image.propTypes = {
+  unoptimized: null,
+};
+
+Image.defaultProps = {
+  unoptimized: true,
+};
