--- front/components/button/Button.stories.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-import { ComponentStory, ComponentMeta } from '@storybook/react';
-import Button from '@components/button/Button';
-
-
-export default {
-    title: 'Button',
-    component: Button,
-} as ComponentMeta<typeof Button>;
-
-const Template: ComponentStory<typeof Button> = (args) => <Button {...args} />;
-
-
-
-export const Default = Template.bind({});
-Default.args = {
-    label:"テストラベル"
-};
