--- front/components/button/Button.tsx
@@ -1,8 +1,0 @@
-import BaseButton, { BaseButtonProps } from '@components/button/templates/BaseButton'
-
-const Button: React.FC<BaseButtonProps> = (props) => {
-
-    return (<BaseButton {...props} />)
-}
-
-export default Button
