--- front/components/cards/messageCard/MessageCard.stories.tsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import { ComponentStory, ComponentMeta } from '@storybook/react';
-import MessageCard from '@components/cards/messageCard/MessageCard';
-
-export default {
-    title: 'MessageCard',
-    component: MessageCard,
-} as ComponentMeta<typeof MessageCard>;
-
-const Template: ComponentStory<typeof MessageCard> = (args) => <MessageCard {...args} />;
-
-
-export const Default = Template.bind({});
-Default.args = {
-    className:"",
-    message:"テストメッセージ"
-};
