+++ front/components/common/button/Button.stories.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import { ComponentStory, ComponentMeta } from '@storybook/react';
+import Button from '@components/common/button/Button';
+
+
+export default {
+    title: 'Button',
+    component: Button,
+} as ComponentMeta<typeof Button>;
+
+const Template: ComponentStory<typeof Button> = (args) => <Button {...args} />;
+
+
+
+export const Default = Template.bind({});
+Default.args = {
+    label:"テストラベル"
+};
