+++ front/components/common/button/Button.tsx
@@ -1,0 +1,8 @@
+import BaseButton, { BaseButtonProps } from '@components/common/button/templates/BaseButton'
+
+const Button: React.FC<BaseButtonProps> = (props) => {
+
+    return (<BaseButton {...props} />)
+}
+
+export default Button
