+++ front/components/common/cards/errorCard/ErrorCard.tsx
