+++ front/components/common/cards/templates/BaseCard.tsx
