--- front/components/inputForm/email/Email.stories.tsx
@@ -1,23 +1,0 @@
-import React from 'react';
-import { ComponentStory, ComponentMeta } from '@storybook/react';
-import Email from '@components/inputForm/email/Email';
-
-
-export default {
-    title: 'Email',
-    component: Email,
-} as ComponentMeta<typeof Email>;
-
-const Template: ComponentStory<typeof Email> = (args) => <Email {...args} />;
-
-export const Default = Template.bind({});
-Default.args = {
-    id: "checkbox",
-};
-
-export const Error = Template.bind({});
-Error.args = {
-    id: "checkbox",
-    error:true,
-    errorMessage:"不正なメールアドレスです"
-};
