+++ front/components/mygame/table/GameListTable.stories.tsx
@@ -1,0 +1,26 @@
+import React from 'react';
+import { ComponentStory, ComponentMeta } from '@storybook/react';
+import GameListTable from '@components/mygame/table/GameListTable';
+import sampleImage from '@assets/images/sampleImage.jpg'
+import testImage from '@assets/images/test.png'
+
+export default {
+    title: 'GameListTable',
+    component: GameListTable,
+} as ComponentMeta<typeof GameListTable>;
+
+const Template: ComponentStory<typeof GameListTable> = (args) => <GameListTable {...args} />;
+
+export const Default = Template.bind({});
+Default.args = {
+    gameList: [
+        { title: "荒廃のマリス", image: sampleImage.src, date: new Date() },
+        { title: "星空のマリス", image: testImage.src, date: new Date() },
+        { title: "九頭竜館の殺人", image: testImage.src, date: new Date() }
+    ]
+};
+
+export const NoGame = Template.bind({}); 
+NoGame.args = {
+    gameList: null
+};
