+++ front/package.json
@@ -24,6 +24,7 @@
     "react-dom": "18.2.0",
     "react-hook-form": "^7.37.0",
     "react-icons": "^4.6.0",
+    "usehooks-ts": "^2.9.1",
     "yup": "^0.32.11"
   },
   "devDependencies": {
