+++ front/pages/signin.tsx
@@ -10,7 +10,7 @@ import useAuth from "@hooks/useAuth"
 import ErrorCard from "@components/cards/errorCard/ErrorCard"
 import HyperLink from "@components/hypyerLink/HyperLink"
 import Divider from "@components/divider/Divider"
-import Button from "@components/button/Button"
+import Button from "@components/common/button/Button"
 import Text from "@components/inputForm/text/Text"
 import CheckBox from "@components/inputForm/checkbox/CheckBox"
 import Password from "@components/inputForm/password/Password"
