+++ front/tsconfig.json
@@ -21,8 +21,7 @@
       "@hooks/*":["./hooks/*"],
       "@graphql/*":["./graphql/*"],
       "@state/*":["./state/*"],
-      "@storybook/*":["./storybook/*"],
-      "@styles/*":["./styles/*"]
+      "@utils/*":["./utils/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
