+++ front/utils/dateUtil.ts
@@ -1,0 +1,4 @@
+export const dateToString = (date: Date | null): string => {
+    if (!date) return ""
+    return `${date.getFullYear()}/${date.getMonth() + 1}/${date.getDate()}`
+}
